+++ components/core-components/blockquote.ts
@@ -11,7 +11,11 @@ import {
 } from '@bangle.dev/pm';
 import { filter, findParentNodeOfType, insertEmpty } from '@bangle.dev/utils';
 import type { MarkdownSerializerState } from 'prosemirror-markdown';
-import { copyEmptyCommand, cutEmptyCommand, moveNode } from './core-commands';
+import {
+  copyEmptyCommand,
+  cutEmptyCommand,
+  moveNode,
+} from '@bangle.dev/pm-commands';
 
 export const spec = specFactory;
 export const plugins = pluginsFactory;