+++ components/core-components/bullet-list.ts
@@ -11,7 +11,7 @@ import {
 } from '@bangle.dev/pm';
 import type Token from 'markdown-it/lib/token';
 import type { MarkdownSerializerState } from 'prosemirror-markdown';
-import { parentHasDirectParentOfType } from './core-commands';
+import { parentHasDirectParentOfType } from '@bangle.dev/pm-commands';
 import { toggleList } from './list-item/commands';
 import { listIsTight } from './list-item/list-is-tight';
 import {