+++ components/core-components/list-item/list-item-component.ts
@@ -15,13 +15,13 @@ import {
 } from '@bangle.dev/utils';
 import type Token from 'markdown-it/lib/token';
 import type { MarkdownSerializerState } from 'prosemirror-markdown';
-import type { MoveDirection } from '../core-commands';
+import type { MoveDirection } from '@bangle.dev/pm-commands';
 import {
   copyEmptyCommand,
   cutEmptyCommand,
   moveNode,
   parentHasDirectParentOfType,
-} from '../core-commands';
+} from '@bangle.dev/pm-commands';
 import {
   backspaceKeyCommand,
   enterKeyCommand,