+++ lib/core/utils/utils.ts
@@ -1,6 +1,0 @@
-export { default as browser } from './browser';
-export * from './logging';
-export * from './mark-input-rule';
-export * from './mark-paste-rule';
-export * from './plugin-key-store';
-export * from './plugin-loader';